refactor(desktop): split app shell daily review actions
